+++ axis_camera/axis_camera/camera_info_manager.py
