+++ axis_camera/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/axis_camera
+[install]
+install_scripts=$base/lib/axis_camera
