+++ axis_camera/test/test_copyright.py
