+++ axis_camera/test/test_flake8.py
@@ -1,4 +1,4 @@
-# Copyright 2015 Open Source Robotics Foundation, Inc.
+# Copyright 2017 Open Source Robotics Foundation, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,13 +12,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from ament_pep257.main import main
+from ament_flake8.main import main_with_errors
 import pytest
 
 
+@pytest.mark.flake8
 @pytest.mark.linter
-@pytest.mark.pep257
-def test_pep257():
-    rc = main(argv=['--exclude', './axis_camera_ros2/camera_info_manager.py',
-                    './axis_camera_ros2/cam_info_utils.py'])
-    assert rc == 0, 'Found code style errors / warnings'
+def test_flake8():
+    rc, errors = main_with_errors(argv=['--exclude', './axis_camera/camera_info_manager.py',
+                                        './axis_camera/cam_info_utils.py'])
+    assert rc == 0, \
+        'Found %d code style errors / warnings:\n' % len(errors) + \
+        '\n'.join(errors)
