--- axis_camera_ros2/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/axis_camera_ros2
-[install]
-install_scripts=$base/lib/axis_camera_ros2
